Add JWT code

+++ resources/item.py
@@ -1,21 +1,22 @@
 from flask_restful import Resource, reqparse
-from flask_jwt import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt_claims, get_jwt_identity, jwt_optional,fresh_jwt_required
 from models.item import ItemModel
+
 
 class Item(Resource):
     parser = reqparse.RequestParser()
     parser.add_argument('price',
-        type=float,
-        required=True,
-        help="This field cannot be left blank!"
-    )
+                        type=float,
+                        required=True,
+                        help="This field cannot be left blank!"
+                        )
     parser.add_argument('store_id',
-        type=int,
-        required=True,
-        help="Every item needs a store id."
-    )
+                        type=int,
+                        required=True,
+                        help="Every item needs a store_id."
+                        )
 
-    @jwt_required()
+    @jwt_required  # No longer needs brackets
     def get(self, name):
         item = ItemModel.find_by_name(name)
         if item:
@@ -36,23 +37,26 @@ class Item(Resource):
             return {"message": "An error occurred inserting the item."}, 500
 
         return item.json(), 201
-
+    @jwt_required
     def delete(self, name):
+        claims = get_jwt_claims()
+        if not claims['is_admin']:
+            return {'message':'Admin privilege required'},401
         item = ItemModel.find_by_name(name)
         if item:
             item.delete_from_db()
-
-        return {'message': 'Item deleted'}
+            return {'message': 'Item deleted.'}
+        return {'message': 'Item not found.'}, 404
 
     def put(self, name):
         data = Item.parser.parse_args()
 
         item = ItemModel.find_by_name(name)
 
-        if item is None:
-            item = ItemModel(name, **data)
-        else:
+        if item:
             item.price = data['price']
+        else:
+            item = ItemModel(name, **data)
 
         item.save_to_db()
 
@@ -60,5 +64,13 @@ class Item(Resource):
 
 
 class ItemList(Resource):
+    @jwt_optional
     def get(self):
-        return {'items': [x.json() for x in ItemModel.query.all()]}
+        user_id = get_jwt_identity()
+        items = [item.json() for item in ItemModel.find_all()]
+        if user_id:
+            return {'items': items},200
+        return {
+            'items': [item['name'] for item in items],
+            'message': 'More data available if you log in'
+        },200
